refactor: Room 데이터베이스를 하나로 통합
